[Update] テスト修正。mockを別ファイルに分離した。
